redesign for more understandable names
